backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
reportlab==4.0.9
python-dotenv>=1.0
requests>=2.31
